Create Swapping index position
A company wants to transmit data over the telephone, but is concerned that its phones may be taped. Develop a program that will encrypt the company's data so that the data may be transmitted more securely. All of its data is transmitted as a four-digit integer. The program must read a four-digit integer entered by the user and encrypt it as follows
Step 1: Replace each digit with the result of adding 5 to the digit and getting the remainder after dividing the new value by 10
Step 2: Swap the fist digit with the third
Step 3: Swap the second digit with the fourth
Step 4: Print the encrypted integer

Explanation:
Input is 4537
(4+5)%10,(5+5)%10,(3+5)%10,(7+5)%10
9082
Swap position 1 with position 3
8092
Swap position 2 with position 4
8290
Output is 8290

TEST CASE 1:
6985

OUTPUT:
3014

TEST CASE 2:
9854

OUTPUT:
943