Create Extract the number from given string
Benny once had trouble finding the number in a string. There are various numbers in the string and you need to extract the string. 
One catch is, you need to find only those number which do not contain 9. For example, if the string contains "hello this is alpha 5051
and 9475". You will extract 5051 and not 9475. Help find the numnber

TEST CASE 1:
1

OUTPUT:
Hi5 Welcome567 batch 2019
5 567 

TEST CASE 2:
3
This is alpha 5057 and 97
GSLV F11 was launched in december 19, 2018
GSLV is a satellite

OUTPUT:
5057   
11      2018 
-1