Create Find the missing number
You are given a list n-1 integer and these integers are in the range of 1 to n. There are no duplicate in list. one of 
the integers is missing in the list. Write an efficient code to find the missing integer. If no number is missing print -1.

TEST CASE 1:
7
1 2 4 6 3 7 8

OUTPUT:
5

TEST CSAE 2:
7
7 6 5 4 3 2 1

OUTPUT:
-1